Join, Apply*: Implement with Compose
Implements Join and all ApplyN operations with the new Compose function
instead of the applyNStep helper we previously had.
